Create Bonus Score
An integer is given - starting number of points. Bonus points are accrued on it according to the rules described below. Write a program that calculates the bonus points that the number receives and the total number of points (the number + the bonus).

• If the number is up to and including 100, the bonus points are 5.
• If the number is greater than 100, the bonus points are 20% of the number.
• If the number is greater than 1000, the bonus points are 10% of the number.

• Additional bonus points (charged separately from the previous ones):
o For an even number + 1 pt.
o For a number that ends in 5 + 2 pts.