Create Toy Shop
Petya has a children's toy shop. She receives a large order to fulfill. With the money he will earn, he wants to go on a trip.

Toy prices:
• Puzzle - BGN 2.60.
• Talking doll - BGN 3
• Teddy bear - BGN 4.10.
• Mignon - BGN 8.20.
• Truck - BGN 2

When ordering 50 toys or more, the store gives a discount of 25% of the total price. From the money earned, Petya must give 10% for the rent of the shop. To calculate whether the money will be enough for her to go on an excursion.

6 lines are read from the console:
1. Excursion price - real number in the interval [1.00 … 10000.00]
2. Number of puzzles - an integer in the range [0… 1000]
3. Number of talking dolls - an integer in the interval [0 … 1000]
4. Number of teddy bears - an integer in the interval [0 … 1000]
5. Number of minions - an integer in the interval [0 … 1000]
6. Number of trucks - an integer in the interval [0 … 1000]

Exit
The console prints:
• If the money is sufficient, the following is printed:
o "Yes! {remaining money} BGN."
• If the money is NOT enough, the following is printed:
o "There is not enough money! BGN is needed."

The result must be formatted to the second decimal place.