Create Godzila vs. Kong
Filming begins on the highly anticipated Godzilla vs. Kong movie. Screenwriter Adam Wingard asks you to write a program to calculate whether the budgeted funds are sufficient to make the movie. Filming will require a certain number of extras, clothing for each extra and set.

It is known that:
• The film set is worth 10% of the budget.
• With over 150 extras, there is a 10% discount on clothes.

3 lines are read from the console:
Line 1. Film budget - real number in interval [1.00 … 1000000.00]
Line 2. Number of extras – an integer in the interval [1 … 500]
Line 3. Price for clothing of one extra - a real number in the interval [1.00 … 1000.00]

Exit
Two lines should be printed to the console:
• If the money for decor and clothes is more than the budget:
o "Not enough money!"
o "Wingard needs {the money short for the film} lev more."
• If the money for decor and clothes is less than or equal to the budget:
o "Action!"
o "Wingard starts filming with {remaining money} lev."
The result must be formatted to the second decimal place.