Create World Swimming Record
Ivan decides to improve the world record in long-distance swimming. Enter on the console the record in seconds that Ivan must break, the distance in meters that he must swim, and the time in seconds that he swims 1 m. Write a program that calculates whether he has completed the task, meaning that: the water resistance slows it down every 15m by 12.5 seconds. When calculating how many times Ivan will slow down due to water resistance, the result should be rounded down to the nearest whole number. Calculate the time in seconds for Ivan to swim the distance and the gap to the world record.

3 lines are read from the console:
1. The record in seconds – a real number in the interval [0.00 … 100000.00]
2. The distance in meters – a real number in the interval [0.00 … 100000.00]
3. The time in seconds to swim a distance of 1 m - a real number in the interval [0.00 … 1000.00]

Exit
Printing to the console depends on the result:
• If Ivan has improved the world record (his time is less than the record) we print: "Yes, he did! The new world record is {Ivan's time} seconds."
• If he did NOT beat the record (his time is greater than or equal to the record) we print:  "No, he failed! He was {missing seconds} seconds slower."

The result must be formatted to the second decimal place.