Create Shopping
Peter wants to buy N video cards, M processors, and P amount of RAM. If the number of graphics cards is greater than the number of processors, you get a 15% discount on the final bill. The following prices apply:

• Video card – BGN 250/pc.
• Processor – 35% of the price of purchased video cards/pcs.
• RAM memory – 10% of the price of purchased video cards/pcs.

To calculate the amount needed to purchase the materials and to calculate whether the budget will reach.

The input consists of four lines:
1. Peter's budget - a real number in the interval [0.0…100000.0]
2. The number of video cards - an integer in the interval [0…100]
3. The number of processors - an integer in the interval [0…100]
4. The number of RAM memory - an integer in the interval [0…100]

Exit
1 line is printed to the console, which should look like this:
• If the budget is sufficient: "You have {remaining budget} BGN left!"
• If the amount exceeds the budget: "The money is not enough! I need another lev!"

Format the result to the second decimal place.