Create Lunch Break
On your lunch break, you want to watch an episode of your favorite series. Your task is to write a program that tells you if you have enough time to watch the episode. During the break you have lunch time and rest time. Lunch time will be 1/8 of the break and recess will be 1/4 of the break.

3 lines are read from the console:
1. Series name - text
2. Episode duration - an integer in the range [10… 90]
3. Rest duration - an integer in the range [10… 120]

Exit
Type one line in the console:
• If there is enough time to watch the episode: "You have enough time to watch {series name} and you have {remaining time} minutes to spare."
• If you don't have enough time: "You don't have enough time to watch {series name}, you need more {necessary time} minutes."

Time to round to the nearest integer.